ioc_aggregator/requirements.txt
requests
pandas
ipwhois
python-whois



